Workaround the selection bug in EditMode.Place.cs
Workaround for the selection bug.
Repro:
- Switch to the Rotation tool
- Select a non-brush transform with the Rotation tool active
- Switch to the Move tool
The scene view will block all interactions (besides dragging the gizmo of the active transform - though the gizmo also won't respond to hovering).

The issue is caused by boundsCenter turning into NaN.
I think the symptoms arise because the Handles calculations within the block I skip end up thinking the bounds are infinitely large - but skipping this chunk works just as well for now.
A true fix would be to prevent boundsCenter from becoming NaN in the first place, but that'll have to wait for another day.